refactor: Rename `lib/plugins/aws/utils/getLambdaLayerArtifactPath.js`
