+++ lib/plugins/aws/deploy/lib/uploadArtifacts.js
@@ -8,7 +8,7 @@ const crypto = require('crypto');
 const limit = require('ext/promise/limit').bind(Promise);
 const filesize = require('filesize');
 const normalizeFiles = require('../../lib/normalize-files');
-const getLambdaLayerArtifactPath = require('../../utils/getLambdaLayerArtifactPath');
+const getLambdaLayerArtifactPath = require('../../utils/get-lambda-layer-artifact-path');
 const ServerlessError = require('../../../../serverless-error');
 const { progress, log } = require('@serverless/utils/log');
