+++ lib/plugins/aws/utils/get-lambda-layer-artifact-path.js
